new and old versions of player_details and visuals exist
